\name{CD4Ids}
\alias{CD4LLids}
\alias{CD4GOTermsFrame}
\alias{CD4GOTermsList}
\docType{data}
\title{Entrez identifiers for CD4-TCells example}
\description{This dataset contains the entrez identifiers \texttt{CD4EntrezIds} and their associated 
GO Terms \texttt{CD4GOTermsFrame} and \texttt{CD4GOTermsList} corresponding to the list 
of differentially expressed genes in a study by Henkel et al.}
\usage{data(CD4Ids)}
\source{
Hengel, R.L. and Thaker, V. and Pavlick, M.V. and Metcalf, J.A. and Dennis,
G. Jr. and Yang, J. and Lempicki, R.A. and Sereti, I. and Lane, H.C. (2003).
L-selectin (CD62L) expression distinguishes small resting memory CD4+ T
cells that preferentially respond to recall antigen. J. Immunol., 170, 28-32. (2003)}
\examples{
data(CD4Ids)
}
\keyword{datasets}
