% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustKidneyMF2.R
\docType{data}
\name{clustKidneyMF2}
\alias{clustKidneyMF2}
\title{Ready 2 cluster equivalence distance matrix obtained from the analysis of the "Kidney Dataset" at level 2 of the MF ontology}
\format{An object of class \code{equivClust} (inherits from \code{hclust}) of length 7.}
\usage{
clustKidneyMF2
}
\description{
An object of class "list" with the information outputted from of the analysis performed by function "profileEquiv_topDown2"
}
\keyword{datasets}
