\name{drosophila}
\alias{drosophilaIds}
\alias{ostrinIds}
\alias{michaudIds}
\docType{data}
\title{Entrez identifiers for genes related with an eye mutation in drosophila}
\description{Entrez identifiers for genes related with an eye mutation in
drosophila.
  \describe{
    \item{\code{ostrinIds}}{List of genes in Entrez, generated by Ostrin et al.}
    \item{\code{michaudIds}}{List of genes in Entrez, generated by Michaud et
      al.} 
    \item{\code{drosophilaIds}}{List of Drosophila genes in Entrez.}
}
}

\usage{data(drosophila)}
\format{Each dataset is a character vector with a different number of elements which (should) correspond to valid Entrez identifiers}
\examples{
data(drosophila)
}
\keyword{datasets}
