% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equivClust2pdf.R
\name{equivClust2pdf}
\alias{equivClust2pdf}
\title{Save the graphical representation of objects of class \code{equivClust} or \code{iterEquivClust}
as pdf files.}
\usage{
equivClust2pdf(x, ...)
}
\arguments{
\item{x}{an object of class \code{equivClust} or \code{iterEquivClust}}

\item{...}{additional arguments to \code{hclust}}

\item{jobName}{character, main plot title and file name (it should be correct as a file name!)}

\item{ylab}{character, label of the plot vertical axis}
}
\description{
Save the graphical representation of objects of class \code{equivClust} or \code{iterEquivClust}
as pdf files.
}
\examples{
data(clustKidneyMF2)
equivClust2pdf(clustKidneyMF2)
# And then open file "Equivalence cluster_MF_2_complete.pdf"...
equivClust2pdf(clustKidneyMF2, jobName = "")
# And then open file "Method 'complete' dendrogram for level 2 of GO ontology MF.pdf"...
}
