\name{goProfiles-package}
\alias{goProfiles-package}
\alias{goProfiles}
\docType{package}
\title{
Performs Gene Ontology based analysis using Functional Profiles.
}
\description{
Performs Gene Ontology based analysis for gene sets or other type of biological identifiers which can be annotated in the Gene Ontology.}
\details{
\tabular{ll}{
Package:\tab goProfiles \cr
Type: \tab Package \cr
Version: \tab 1.0 \cr
Date: \tab 2007-06-12 \cr
License:\tab GPL \cr
}
}
\author{
Alex Sanchez and Jordi Ocana
}
\references{
Sanchez-Pla, A., Salicru M. and J.Ocana.
Statistical methods for the analysis of high-throughput data based on
functional profiles derived from the gene ontology.
Journal of Statistical Planning and Inference, 2007.
}
\keyword{ package }
\seealso{
goTools, GOstat, ontoTools, and other Bioconductor packages for GO based analysis}
