\name{hugoIds}
\alias{hugoIds}
\docType{data}
\title{Entrez Identifiers obtained from the Human Genome Organization}
\description{Entrez identifiers obtained from the Human Genome
  Organization. They correspond to the column named 'Entrez Gene Id
  (mapped)' in the 'All data' table in the Hugo Genome Nomenclature web site
  (http://www.genenames.org/index.html)}
\references{\url{http://www.genenames.org/cgi-bin/hgnc_downloads.cgi}}
\usage{data(hugoIds)}
\examples{
data(hugoIds)}
\keyword{datasets}
