% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidneyGeneLists.R
\docType{data}
\name{kidneyGeneLists}
\alias{kidneyGeneLists}
\title{Gene-lists related to kidney transplantation rejection}
\format{An object of class "list" with 5 character vectors:}
\source{
\url{https://www.ualberta.ca/medicine/institutes-centres-groups/atagc/research/gene-list}
}
\usage{
kidneyGeneLists
}
\description{
An object of class "list" containing a selected subset of 5 gene-lists related to kidney transplantation rejection,
described generically as "PBTS" (Patogenic Based Transcript Sets).
Each gene-list is a character vector containing "Entrez" identifiers (integer numbers) for all the genes it contains.
}
\keyword{datasets}
