\name{sampleProfiles}
\alias{sampleProfiles}
\alias{basicWelsh01} 
\alias{basicSingh01} 
\alias{basicCD4} 
\alias{basicHugo} 
\alias{expandedWelsh01} 
\alias{expandedSingh01} 
\alias{commonExpandedWelsh01Singh01} 
\alias{expandedCD4} 
\alias{expandedHugo}
\docType{data}
\title{Sample Profiles computed on some datasets in the package}
\description{This dataset contains some basic and expanded sample profiles computed on some of the datasets in the package.
 They are intended to be used in examples of profile manipulation, without having to re=compute the profiles each time the example is checked.
 The prefix of the object indicates if it is a 'basic' or an 'expanded' profile, 
 whereas the name can be clearly associated to some of the datasets available in the package.
\item \code{basicWelsh01} Basic (contracted) profile made from \code{Welsh01EntrezIds} dataset with default parameters.
\item \code{basicSingh01} Basic (contracted) profile made from \code{Singh01EntrezIds} dataset with default parameters.
\item \code{basicCD4} Basic (contracted) profile made from \code{CD4LLids} dataset with default parameters.
\item \code{basicHugo}  Basic (contracted) profile made from \code{hugoIds} dataset with default parameters.
\item \code{expandedWelsh01} Expanded profile made from \code{Welsh01EntrezIds} dataset with default parameters.
\item \code{expandedSingh01} Expanded profile made from \code{Singh01EntrezIds} dataset with default parameters.
\item \code{commonExpandedWelsh01Singh01} Expanded profile made from common genes in \code{Welsh01EntrezIds} and \code{Singh01EntrezIds} dataset with default parameters.
\item \code{expandedCD4} Expanded profile made from \code{CD4LLids} dataset with default parameters.
\item \code{expandedHugo} Expanded profile made from \code{hugoIds} dataset with default parameters.
}
\usage{data(sampleProfiles)}

\examples{
data(sampleProfiles)
}
\keyword{datasets}
